Renaming the ids from the activity_movie_detailed.xml and make all of the fragments scrollable
